+++ lib/oli/delivery/sections/section.ex
@@ -45,6 +45,7 @@ defmodule Oli.Delivery.Sections.Section do
     field(:nrps_context_memberships_url, :string)
 
     field(:resource_gating_index, :map, default: %{}, null: false)
+    field(:previous_next_index, :map, default: nil, null: true)
 
     belongs_to(:lti_1p3_deployment, Lti_1p3.DataProviders.EctoProvider.Deployment,
       foreign_key: :lti_1p3_deployment_id
@@ -108,6 +109,7 @@ defmodule Oli.Delivery.Sections.Section do
       :nrps_enabled,
       :nrps_context_memberships_url,
       :resource_gating_index,
+      :previous_next_index,
       :lti_1p3_deployment_id,
       :institution_id,
       :base_project_id,