+++ lib/oli_web/plugs/set_user.ex
@@ -1,7 +1,7 @@
 defmodule Oli.Plugs.SetCurrentUser do
   import Plug.Conn
 
-  alias Oli.Accounts.Author
+  alias Oli.Accounts
   alias Oli.Accounts.User
   alias Oli.Repo
 
@@ -20,14 +20,18 @@ defmodule Oli.Plugs.SetCurrentUser do
 
     if author = Pow.Plug.current_user(conn, pow_config) do
       cond do
-        current_author = Repo.get(Author, author.id) ->
+        current_author = Accounts.get_author_with_community_admin_count(author.id) ->
           conn
           |> put_session(:current_author_id, current_author.id)
+          |> put_session(:is_community_admin, current_author.community_admin_count > 0)
+          |> put_session(:is_system_admin, Accounts.is_admin?(current_author))
           |> assign(:current_author, current_author)
 
         true ->
           conn
           |> delete_session(:current_author_id)
+          |> delete_session(:is_community_admin)
+          |> delete_session(:is_system_admin)
           |> assign(:current_author, nil)
       end
     else